Switching fully to PyTorch
