+++ src/SMPredict.py
@@ -11,7 +11,7 @@ class SMPredict:
 
         # Setting the model
         self.model = torch.nn.Linear(self.inp_num, self.out_num)
-        torch.nn.init.xavier_uniform(self.model.weight)
+        torch.nn.init.xavier_uniform_(self.model.weight)
         self.optimizer = torch.optim.Adam(self.model.parameters(), lr=self.lr)
         self.loss = torch.nn.BCEWithLogitsLoss()
 
@@ -20,22 +20,22 @@ class SMPredict:
     def update(self, patterns, labels):
         self.optimizer.zero_grad()
         output = self.model(torch.tensor(patterns).float())
-        loss = self.loss(output, torch.tensor(labels).unsqueeze(-1))
+        loss = self.loss(output, torch.tensor(labels).float())
         loss.backward()
         self.optimizer.step()
         return loss
 
     def get_weights(self):
         # return self.out_layer.get_weights()[0]
-        return self.model.weight
+        return self.model.weight.detach().cpu().numpy()
 
     def set_weights(self, weights):
         with torch.no_grad():
-            self.model.copy_(weights)
+            self.model.weight.copy_(torch.tensor(weights, dtype=torch.float))
 
     def spread(self, inp):
         assert len(inp.shape) == 2
-        comp = self.model(torch.tensor(inp).float()).detach().numpy()
+        comp = self.model(torch.tensor(inp, dtype=torch.float)).detach().cpu().numpy()
         comp = 2*np.maximum(0, comp - 0.5)
         return comp
 
@@ -51,6 +51,7 @@ if __name__ == "__main__":
     labels[patterns_size//2:] = 1
     patterns = np.vstack([labels, 1-labels]).T \
             + 0.01*np.random.randn(patterns_size, 2)
+    labels = labels[:, None]
     predict = SMPredict(inp_num, out_num)
 
     idcs = np.arange(patterns_size)
